double the steps taken statistic
Crew have two feet so it should increment twice anyway. Also the stats are
surprisingly low and this is a way to make the numbers look bigger.
I'm only incrementing steps taken twice each time the crewmember moves into a
new grid square. But instead, I could actually estimate the number of animation
frames that the crewmember takes on average to move from one square to the
next...